+++ src/service.rs
@@ -1,8 +1,7 @@
-use std::collections::HashMap;
 use std::path::PathBuf;
 use crate::{m3u_processor, utils, xtream_processor};
-use crate::config::ConfigInput;
-use crate::m3u::PlaylistGroup;
+use crate::config::{Config, ConfigInput};
+use crate::m3u::{PlaylistGroup, XtreamCluster};
 
 fn prepare_file_path(input: &ConfigInput, working_dir: &String, action: &str, verbose: bool) -> Option<PathBuf> {
     let persist_file: Option<PathBuf> =
@@ -18,11 +17,11 @@ fn prepare_file_path(input: &ConfigInput, working_dir: &String, action: &str, ve
 }
 
 
-pub(crate) fn get_m3u_playlist(input: &ConfigInput, working_dir: &String, verbose: bool) -> Option<Vec<PlaylistGroup>> {
+pub(crate) fn get_m3u_playlist(cfg: &Config, input: &ConfigInput, working_dir: &String, verbose: bool) -> Option<Vec<PlaylistGroup>> {
     let url = input.url.as_str();
     let file_path = prepare_file_path(input, working_dir, "", verbose);
     let lines: Option<Vec<String>> = utils::get_input_content(working_dir, url, file_path, verbose);
-    lines.map(|l| m3u_processor::process(&l))
+    lines.map(|l| m3u_processor::process(cfg,&l))
 }
 
 
@@ -31,11 +30,11 @@ pub(crate) fn get_xtream_playlist(input: &ConfigInput, working_dir: &String, ver
     let base_url = format!("{}/player_api.php?username={}&password={}", input.url, input.username, input.password);
     let stream_base_url = format!("{}/{}/{}", input.url, input.username, input.password);
 
-    let actions = HashMap::<String, String>::from([
-        (String::from("get_live_categories"), String::from("get_live_streams")),
-        (String::from("get_vod_categories"), String::from("get_vod_streams")),
-        (String::from("get_series_categories"), String::from("get_series"))]);
-    for (category, stream) in &actions {
+    let actions = [
+        (XtreamCluster::LIVE, String::from("get_live_categories"), String::from("get_live_streams")),
+        (XtreamCluster::VIDEO, String::from("get_vod_categories"), String::from("get_vod_streams")),
+        (XtreamCluster::SERIES, String::from("get_series_categories"), String::from("get_series"))];
+    for (xtream_cluster, category, stream) in &actions {
         let category_url =  format!("{}&action={}", base_url, category);
         let stream_url =  format!("{}&action={}", base_url, stream);
         let category_file_path = prepare_file_path(input, working_dir, format!("{}_", category).as_str(), verbose);
@@ -43,7 +42,7 @@ pub(crate) fn get_xtream_playlist(input: &ConfigInput, working_dir: &String, ver
 
         let category_content: Option<serde_json::Value> = utils::get_input_json_content(input, &category_url, category_file_path, verbose);
         let stream_content: Option<serde_json::Value> = utils::get_input_json_content(input, &stream_url, stream_file_path, verbose);
-        let mut sub_playlist = xtream_processor::process(category_content, stream_content, &stream_base_url);
+        let mut sub_playlist = xtream_processor::process(xtream_cluster, category_content, stream_content, &stream_base_url);
         while let Some(group) = sub_playlist.pop() {
             playlist.push(group);
         }